+++ example/TestLinkExample.py
@@ -264,20 +264,25 @@ print("createTestCase", newTestCase)
 newTestCaseID_AA = newTestCase[0]['id']
 print("New Test Case '%s' - id: %s" % (NEWTESTCASE_AA, newTestCaseID_AA))              
 
-#Creates the test case TC_B  with state rework
+#Creates the test case TC_B  with state rework - in wrong test suite A
 myTestLink.initStep("Step action 1", "Step result 1", AUTOMATED)
 myTestLink.appendStep("Step action 2", "Step result 2", AUTOMATED)
 myTestLink.appendStep("Step action 3", "Step result 3", AUTOMATED)
 myTestLink.appendStep("Step action 4", "Step result 4", AUTOMATED)
 myTestLink.appendStep("Step action 5", "Step result 5", AUTOMATED)
      
-newTestCase = myTestLink.createTestCase(NEWTESTCASE_B, newTestSuiteID_B, 
+newTestCase = myTestLink.createTestCase(NEWTESTCASE_B, newTestSuiteID_A, 
           newProjectID, myTestUserName, "This is the summary of the Test Case B", 
           preconditions='these are the preconditions', executiontype=AUTOMATED, 
           status=REWORK, estimatedexecduration=0.5)   
-print("createTestCase", newTestCase)
+print("createTestCase TC-B in TS-A", newTestCase)
 newTestCaseID_B = newTestCase[0]['id']
 print("New Test Case '%s' - id: %s" % (NEWTESTCASE_B, newTestCaseID_B))               
+
+# Move test case TC_B to correct test suite B
+tc_b_full_ext_id = myTestLink.getTestCase(newTestCaseID_B)[0]['full_tc_external_id']
+response = myTestLink.setTestCaseTestSuite(tc_b_full_ext_id, newTestSuiteID_B)
+print("setTestCaseTestSuite TC-B to TS-B" , response)
   
 # Add  test cases to test plan - we need the full external id !
 # for every test case version 1 is used