+++ test/utest-online/testlinkapi_online_test.py
@@ -472,6 +472,12 @@ class TestLinkAPIOnlineTestCase(unittest.TestCase):
         with self.assertRaisesRegex(TLResponseError, '7000.*40000712'):
             self.client.getReqCoverage(
                             '40000712 project', '40000721 req')
+
+    def test_setTestCaseTestSuite_unknownID(self):
+        with self.assertRaisesRegex(TLResponseError, '5040.*TC-40000712'):
+            self.client.setTestCaseTestSuite(
+                            'TC-40000712', '40000713 suite')
+         
         
 if __name__ == "__main__":
     #import sys;sys.argv = ['', 'Test.testName']